Completed BASIS integration and tests.
