+++ tests/test_basis.py
@@ -1,11 +1,29 @@
 """Test the BASIS devices/driver respond with correct data."""
 from random import uniform
+from unittest import mock
 
 import pytest
 
 from alicat.basis import BASISController, BASISMeter
+from alicat.mock import BASISClient
 
 ADDRESS = "COM16" # tests requite unit: A, baud: 38400
+
+@pytest.fixture(scope='module', autouse=True)
+async def patch_serial_client(request):
+    """Replace the serial client with our mock."""
+    if request.config.getoption("--hardware"):
+        async with BASISMeter(ADDRESS) as device:
+            res = await device._write_and_read('A')
+            if not res:
+                pytest.exit("Device not found. Ensure device is connected " \
+                "with unit id A and BAUD rate 38400 on the correct COM port.")
+        yield
+        async with BASISMeter(ADDRESS) as device:
+            await device._write_and_read('AS 0')
+    else:
+        with mock.patch('alicat.basis.SerialClient', BASISClient):
+            yield
 
 @pytest.mark.skip ## skip
 @pytest.mark.parametrize('cls', [BASISMeter])  # Fixme: fix FlowMeter
